Moving tests to a better class
