+++ pipeline-model-definition/src/test/java/org/jenkinsci/plugins/pipeline/modeldefinition/BuildConditionResponderTest.java
@@ -23,12 +23,23 @@
  */
 package org.jenkinsci.plugins.pipeline.modeldefinition;
 
+import com.gargoylesoftware.htmlunit.html.HtmlPage;
 import hudson.model.Result;
+import hudson.model.queue.QueueTaskFuture;
+import jenkins.model.Jenkins;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
+import org.jenkinsci.plugins.workflow.cps.CpsFlowExecution;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
+import org.jenkinsci.plugins.workflow.support.steps.input.InputAction;
+import org.jenkinsci.plugins.workflow.support.steps.input.InputStepExecution;
+import org.jenkinsci.plugins.workflow.test.steps.SemaphoreStep;
 import org.junit.Test;
 import org.jvnet.hudson.test.Issue;
+import org.jvnet.hudson.test.JenkinsRule;
+import org.jvnet.hudson.test.MockAuthorizationStrategy;
+
+import static org.junit.Assert.assertEquals;
 
 /**
  * @author Andrew Bayer
@@ -98,5 +109,90 @@ public class BuildConditionResponderTest extends AbstractModelDefTest {
                 .logContains("[Pipeline] { (foo)", "hello")
                 .inLogInOrder("I AM ALWAYS", "I CHANGED", "I SUCCEEDED")
                 .go();
+    }
+
+    @Issue("JENKINS-43339")
+    @Test
+    public void notBuiltFlowInterruptedException() throws Exception {
+        WorkflowJob job = j.jenkins.createProject(WorkflowJob.class, "milestone-flow-interrupted");
+        job.setDefinition(new CpsFlowDefinition("" +
+                "pipeline {\n" +
+                "  agent none\n" +
+                "  stages {\n" +
+                "    stage('milestones') {\n" +
+                "      steps {\n" +
+                "        milestone(1)\n" +
+                "        semaphore 'wait'\n" +
+                "        milestone(2)\n" +
+                "      }\n" +
+                "      post {\n" +
+                "        notBuilt {\n" +
+                "          echo 'Job not built due to milestone'\n" +
+                "        }\n" +
+                "      }\n" +
+                "    }\n" +
+                "  }\n" +
+                "}\n", true));
+
+        WorkflowRun run1 = job.scheduleBuild2(0).waitForStart();
+        SemaphoreStep.waitForStart("wait/1", run1);
+        WorkflowRun run2 = job.scheduleBuild2(0).waitForStart();
+        SemaphoreStep.waitForStart("wait/2", run2);
+
+        SemaphoreStep.success("wait/2", null);
+
+        j.waitForCompletion(run2);
+
+        j.assertBuildStatus(Result.NOT_BUILT, j.waitForCompletion(run1));
+
+        j.assertLogContains("Job not built due to milestone", run1);
+    }
+
+    @Issue("JENKINS-43339")
+    @Test
+    public void abortedFlowInterruptedException() throws Exception {
+        j.jenkins.setSecurityRealm(j.createDummySecurityRealm());
+        j.jenkins.setAuthorizationStrategy(new MockAuthorizationStrategy().grant(Jenkins.ADMINISTER).everywhere().to("ops"));
+        WorkflowJob job = j.jenkins.createProject(WorkflowJob.class, "input-flow-interrupted");
+        job.setDefinition(new CpsFlowDefinition("" +
+                "pipeline {\n" +
+                "  agent none\n" +
+                "  stages {\n" +
+                "    stage('input') {\n" +
+                "      steps {\n" +
+                "        input(id: 'InputX', message: 'OK?', ok: 'Yes', submitter: 'ops')\n" +
+                "      }\n" +
+                "      post {\n" +
+                "        aborted {\n" +
+                "          echo 'Job aborted due to input'\n" +
+                "        }\n" +
+                "      }\n" +
+                "    }\n" +
+                "  }\n" +
+                "}\n", true));
+
+        QueueTaskFuture<WorkflowRun> queueTaskFuture = job.scheduleBuild2(0);
+        WorkflowRun run = queueTaskFuture.getStartCondition().get();
+        CpsFlowExecution execution = (CpsFlowExecution) run.getExecutionPromise().get();
+
+        while (run.getAction(InputAction.class) == null) {
+            execution.waitForSuspension();
+        }
+
+        JenkinsRule.WebClient webClient = j.createWebClient();
+
+        webClient.login("ops");
+
+        InputAction inputAction = run.getAction(InputAction.class);
+        InputStepExecution is = inputAction.getExecution("InputX");
+        HtmlPage p = webClient.getPage(run, inputAction.getUrlName());
+
+        j.submit(p.getFormByName(is.getId()), "abort");
+        assertEquals(0, inputAction.getExecutions().size());
+        queueTaskFuture.get();
+
+        j.assertBuildStatus(Result.ABORTED, j.waitForCompletion(run));
+
+        j.assertLogContains("Job aborted due to input", run);
     }
 }
